moved gltf loading outside of renderer. working on generic interface to send work to renderer
