+++ src/core/loaders/gltf_loader.cpp
@@ -1,4 +1,4 @@
-#include "vk_loader.h"
+#include "gltf_loader.h"
 #include "fastgltf/core.hpp"
 #include "fastgltf/tools.hpp"
 #include "fastgltf/types.hpp"
@@ -16,33 +16,82 @@
 #include <fastgltf/glm_element_traits.hpp>
 #include <fmt/base.h>
 #include <fmt/format.h>
-#include <fstream>
-#include <iostream>
 #include <string>
 #include <variant>
 #include <vk_backend/vk_sync.h>
 #include <vulkan/vulkan_core.h>
 
-VkShaderModule load_shader_module(VkDevice device, const char* file_path) {
+struct DrawObjUniformData {
+    glm::mat4       local_transform{1.f};
+    VkDeviceAddress vertex_buffer_address;
+};
 
-    std::ifstream         file(file_path, std::ios::ate | std::ios::binary);
-    size_t                file_size = (size_t)file.tellg();
-    std::vector<uint32_t> buffer(file_size / sizeof(uint32_t));
+struct GLTFTexture {
+    AllocatedImage tex;
+    VkSampler      sampler;
+};
 
-    file.seekg(0);
-    file.read((char*)buffer.data(), file_size);
-    file.close();
+struct PBRMetallicRoughness {
+    glm::vec4 color_factors;
+    uint32_t  color_tex_coord{0};
+    uint32_t  metal_rough_tex_coord{0};
+    float     metallic_factor;
+    float     roughness_factor;
+};
 
-    VkShaderModule           shader_module;
-    VkShaderModuleCreateInfo shader_module_ci{};
-    shader_module_ci.sType    = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-    shader_module_ci.codeSize = file_size;
-    shader_module_ci.pCode    = buffer.data();
+struct GLTFMaterial {
+    PBRMetallicRoughness pbr;
+    fastgltf::AlphaMode  alpha_mode;
+};
 
-    VK_CHECK(vkCreateShaderModule(device, &shader_module_ci, nullptr, &shader_module));
+struct GLTFPrimitive {
+    VkDescriptorSet     obj_desc_set;
+    VkDescriptorSet     mat_desc_set;
+    VkBuffer            index_buffer;
+    uint32_t            indices_count;
+    uint32_t            indices_start;
+    uint32_t            mat_idx;
+    fastgltf::AlphaMode alpha_mode;
+};
 
-    return shader_module;
-}
+struct GLTFMesh {
+    std::vector<uint32_t>      indices;
+    std::vector<Vertex>        vertices;
+    glm::mat4                  local_transform{1.f};
+    std::vector<GLTFPrimitive> primitives;
+};
+
+struct GLTFNode {
+    glm::mat4 transform{1.f};
+    uint32_t  mesh_idx;
+};
+
+struct MaterialUniformData {
+    glm::vec4 color_factors;
+    float     metallic_factor;
+    float     roughness_factor;
+};
+
+struct MaterialBuffers {
+    AllocatedBuffer mat_uniform_buffer;
+    AllocatedImage  color_tex;
+    AllocatedImage  metal_rough_tex;
+    /*
+     * this set contains bindings for the above resources
+     * 0. mat_uniform_buffer
+     * 1. color_tex
+     * 2. metal_rough_tex
+     */
+    VkDescriptorSet mat_desc_set;
+};
+
+struct GLTFSceneResources {
+    // currently only holding onto these to delete later
+    std::vector<MeshBuffers>     mesh_buffers;
+    std::vector<MaterialBuffers> material_buffers;
+    std::vector<AllocatedBuffer> draw_obj_uniform_buffers;
+    std::vector<VkSampler>       samplers;
+};
 
 glm::mat4 get_transform_matrix(const fastgltf::Node& node, glm::mat4x4& base) {
     if (const auto* pMatrix = std::get_if<fastgltf::Node::TransformMatrix>(&node.transform)) {
@@ -58,6 +107,61 @@ glm::mat4 get_transform_matrix(const fastgltf::Node& node, glm::mat4x4& base) {
 
     return base;
 }
+
+AllocatedImage download_texture(const VkBackend*   backend,
+                                fastgltf::Asset*   asset,
+                                fastgltf::Texture* gltf_texture) {
+    auto&          image = asset->images[gltf_texture->imageIndex.value()];
+    int            width, height, nr_channels;
+    AllocatedImage new_texture;
+
+    std::visit(fastgltf::visitor{
+                   []([[maybe_unused]] auto& arg) {
+
+                   },
+                   [&](fastgltf::sources::URI& file_path) {
+                       assert(file_path.fileByteOffset == 0);
+                       assert(file_path.uri.isLocalPath());
+
+                       const std::string path(file_path.uri.path().begin(),
+                                              file_path.uri.path().end()); // thanks C++.
+                       uint8_t* data = stbi_load(path.c_str(), &width, &height, &nr_channels, 4);
+                       new_texture =
+                           upload_texture(backend, data, VK_IMAGE_USAGE_SAMPLED_BIT, height, width);
+                       stbi_image_free(data);
+                   },
+                   [&](fastgltf::sources::Array& vector) {
+                       uint8_t* data = stbi_load_from_memory(vector.bytes.data(),
+                                                             static_cast<int>(vector.bytes.size()),
+                                                             &width, &height, &nr_channels, 4);
+                       new_texture =
+                           upload_texture(backend, data, VK_IMAGE_USAGE_SAMPLED_BIT, height, width);
+                       stbi_image_free(data);
+                   },
+                   [&](fastgltf::sources::BufferView& view) {
+                       auto& buffer_view = asset->bufferViews[view.bufferViewIndex];
+                       auto& buffer      = asset->buffers[buffer_view.bufferIndex];
+
+                       std::visit(fastgltf::visitor{
+                                      []([[maybe_unused]] auto& arg) {},
+                                      [&](fastgltf::sources::Array& vector) {
+                                          uint8_t* data = stbi_load_from_memory(
+                                              vector.bytes.data() + buffer_view.byteOffset,
+
+                                              static_cast<int>(buffer_view.byteLength), &width,
+                                              &height, &nr_channels, 4);
+                                          new_texture = upload_texture(backend, data,
+                                                                       VK_IMAGE_USAGE_SAMPLED_BIT,
+                                                                       height, width);
+                                          stbi_image_free(data);
+                                      }},
+                                  buffer.data);
+                   },
+               },
+               image.data);
+
+    return new_texture;
+};
 
 // recursively fills a vector of nodes based on GLTF node tree
 void generate_nodes(std::vector<GLTFNode>& out_node_buf,
@@ -137,7 +241,7 @@ std::vector<VkSampler> get_samplers(VkDevice device, fastgltf::Asset& asset) {
     return samplers;
 }
 
-Scene load_scene(VkBackend* backend, std::filesystem::path path) {
+Entity load_scene(VkBackend* backend, std::filesystem::path path) {
 
     constexpr auto supported_extensions = fastgltf::Extensions::KHR_mesh_quantization |
                                           fastgltf::Extensions::KHR_texture_transform |
@@ -165,7 +269,7 @@ Scene load_scene(VkBackend* backend, std::filesystem::path path) {
     fastgltf::Asset asset;
     asset = std::move(load.get());
 
-    Scene new_scene;
+    GLTFSceneResources new_scene;
 
     new_scene.samplers = get_samplers(backend->device_ctx.logical_device, asset);
 
@@ -190,7 +294,7 @@ Scene load_scene(VkBackend* backend, std::filesystem::path path) {
         {VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, 2}
     };
 
-    init_desc_allocator(&new_scene.mat_desc_allocator, backend->device_ctx.logical_device,
+    init_desc_allocator(&backend->mat_desc_allocator, backend->device_ctx.logical_device,
                         asset.materials.size(), mat_pool_sizes);
 
     std::vector<GLTFMaterial> materials;
@@ -259,7 +363,7 @@ Scene load_scene(VkBackend* backend, std::filesystem::path path) {
         }
 
         new_bufs.mat_desc_set =
-            allocate_desc_set(&new_scene.mat_desc_allocator, backend->device_ctx.logical_device,
+            allocate_desc_set(&backend->mat_desc_allocator, backend->device_ctx.logical_device,
                               backend->mat_desc_set_layout);
 
         // 1. fill in the uniform material buffer
@@ -383,6 +487,7 @@ Scene load_scene(VkBackend* backend, std::filesystem::path path) {
         meshes.push_back(new_mesh);
     }
 
+    new_scene.mesh_buffers.reserve(meshes.size());
     for (auto& mesh : meshes) {
         MeshBuffers new_mesh_bufs = upload_mesh_buffers(backend, mesh.indices, mesh.vertices);
         new_scene.mesh_buffers.push_back(new_mesh_bufs);
@@ -397,12 +502,10 @@ Scene load_scene(VkBackend* backend, std::filesystem::path path) {
     std::vector<PoolSizeRatio> obj_pool_sizes = {
         {VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1}
     };
-    init_desc_allocator(&new_scene.obj_desc_allocator, backend->device_ctx.logical_device,
+    init_desc_allocator(&backend->obj_desc_allocator, backend->device_ctx.logical_device,
                         primitive_count, obj_pool_sizes);
 
     new_scene.draw_obj_uniform_buffers.reserve(primitive_count);
-    new_scene.mesh_buffers.reserve(meshes.size());
-
     std::vector<DrawObject> opaque_draws;
     opaque_draws.reserve(primitive_count);
     std::vector<DrawObject> trans_draws;
@@ -422,12 +525,8 @@ Scene load_scene(VkBackend* backend, std::filesystem::path path) {
 
             AllocatedBuffer obj_uniform_buf;
 
-            // VkDescriptorSet draw_obj_desc_set =
-            // new_scene.obj_desc_allocator.allocate(
-            //     backend.device_ctx.logical_device,
-            //     backend._draw_obj_desc_set_layout);
             VkDescriptorSet draw_obj_desc_set =
-                allocate_desc_set(&new_scene.obj_desc_allocator, backend->device_ctx.logical_device,
+                allocate_desc_set(&backend->obj_desc_allocator, backend->device_ctx.logical_device,
                                   backend->draw_obj_desc_set_layout);
 
             obj_uniform_buf = create_buffer(sizeof(DrawObjUniformData), backend->allocator,
@@ -489,46 +588,73 @@ Scene load_scene(VkBackend* backend, std::filesystem::path path) {
                   }
               });
 
-    new_scene.opaque_objs.resize(opaque_draws.size());
-    new_scene.transparent_objs.resize(trans_draws.size());
+    Entity entity;
+    entity.opaque_objs.resize(opaque_draws.size());
+    entity.transparent_objs.resize(trans_draws.size());
 
-    memcpy(new_scene.opaque_objs.data(), opaque_draws.data(),
+    memcpy(entity.opaque_objs.data(), opaque_draws.data(),
            opaque_draws.size() * sizeof(DrawObject));
-    memcpy(new_scene.transparent_objs.data(), trans_draws.data(),
+    memcpy(entity.transparent_objs.data(), trans_draws.data(),
            trans_draws.size() * sizeof(DrawObject));
 
-    return new_scene;
+    backend->deletion_queue.push_persistant([=]() {
+        DEBUG_PRINT("Destroying Scene");
+
+        VkDevice     device    = backend->device_ctx.logical_device;
+        VmaAllocator allocator = backend->allocator;
+
+        for (auto& sampler : new_scene.samplers) {
+            vkDestroySampler(device, sampler, nullptr);
+        }
+        for (auto mesh : new_scene.mesh_buffers) {
+            destroy_buffer(allocator, &mesh.indices);
+            destroy_buffer(allocator, &mesh.vertices);
+        }
+        for (auto material : new_scene.material_buffers) {
+
+            if (material.color_tex.image != backend->default_texture.image) {
+                destroy_image(device, allocator, material.color_tex);
+            }
+            if (material.metal_rough_tex.image != backend->default_texture.image) {
+                destroy_image(device, allocator, material.metal_rough_tex);
+            }
+            destroy_buffer(allocator, &material.mat_uniform_buffer);
+        }
+
+        for (auto draw_obj_uniform : new_scene.draw_obj_uniform_buffers) {
+            destroy_buffer(allocator, &draw_obj_uniform);
+        }
+    });
+
+    return entity;
 }
 
-void destroy_scene(VkBackend* backend) {
-
-    DEBUG_PRINT("Destroying Scene");
-
-    VkDevice     device    = backend->device_ctx.logical_device;
-    VmaAllocator allocator = backend->allocator;
-
-    deinit_desc_allocator(&backend->scene.mat_desc_allocator, device);
-    deinit_desc_allocator(&backend->scene.obj_desc_allocator, device);
-
-    for (auto& sampler : backend->scene.samplers) {
-        vkDestroySampler(device, sampler, nullptr);
-    }
-    for (auto& mesh : backend->scene.mesh_buffers) {
-        destroy_buffer(allocator, &mesh.indices);
-        destroy_buffer(allocator, &mesh.vertices);
-    }
-    for (auto& material : backend->scene.material_buffers) {
-
-        if (material.color_tex.image != backend->default_texture.image) {
-            destroy_image(device, allocator, material.color_tex);
-        }
-        if (material.metal_rough_tex.image != backend->default_texture.image) {
-            destroy_image(device, allocator, material.metal_rough_tex);
-        }
-        destroy_buffer(allocator, &material.mat_uniform_buffer);
-    }
-
-    for (auto& draw_obj_uniform : backend->scene.draw_obj_uniform_buffers) {
-        destroy_buffer(allocator, &draw_obj_uniform);
-    }
-}
+// void destroy_scene(VkBackend* backend) {
+//
+//     DEBUG_PRINT("Destroying Scene");
+//
+//     VkDevice     device    = backend->device_ctx.logical_device;
+//     VmaAllocator allocator = backend->allocator;
+//
+//     for (auto& sampler : backend->scene.samplers) {
+//         vkDestroySampler(device, sampler, nullptr);
+//     }
+//     for (auto& mesh : backend->scene.mesh_buffers) {
+//         destroy_buffer(allocator, &mesh.indices);
+//         destroy_buffer(allocator, &mesh.vertices);
+//     }
+//     for (auto& material : backend->scene.material_buffers) {
+//
+//         if (material.color_tex.image != backend->default_texture.image) {
+//             destroy_image(device, allocator, material.color_tex);
+//         }
+//         if (material.metal_rough_tex.image != backend->default_texture.image) {
+//             destroy_image(device, allocator, material.metal_rough_tex);
+//         }
+//         destroy_buffer(allocator, &material.mat_uniform_buffer);
+//     }
+//
+//     for (auto& draw_obj_uniform : backend->scene.draw_obj_uniform_buffers) {
+//         destroy_buffer(allocator, &draw_obj_uniform);
+//     }
+// }
