+++ src/vk_backend/resources/vk_image.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <fastgltf/core.hpp>
 #include <vk_backend/vk_types.h>
 
 struct VkBackend;
@@ -20,8 +19,11 @@ AllocatedImage create_image(VkDevice          device,
                             VkFormat          format,
                             uint32_t          samples = 1);
 
-void blit_image(
-    VkCommandBuffer cmd, VkImage src, VkImage dest, VkExtent2D src_extent, VkExtent2D dst_extent);
+void blit_image(VkCommandBuffer cmd,
+                VkImage         src,
+                VkImage         dest,
+                VkExtent2D      src_extent,
+                VkExtent2D      dst_extent);
 
 void destroy_image(VkDevice device, VmaAllocator allocator, AllocatedImage& allocated_image);
 
@@ -45,9 +47,8 @@ VkRenderingAttachmentInfo create_depth_attachment_info(VkImageView         view,
                                                        VkAttachmentStoreOp store_op);
 
 // uploads texture to gpu memory
-AllocatedImage upload_texture(
-    const VkBackend* backend, void* data, VkImageUsageFlags usage, uint32_t height, uint32_t width);
-
-//  downloads texture from external image data (png or jpg) to then load onto the gpu
-AllocatedImage
-download_texture(const VkBackend* backend, fastgltf::Asset* asset, fastgltf::Texture* gltf_texture);
+AllocatedImage upload_texture(const VkBackend*  backend,
+                              void*             data,
+                              VkImageUsageFlags usage,
+                              uint32_t          height,
+                              uint32_t          width);
